Tak 테이블 고유 식별자 타입 변경
Long으로 변경하였습니다.
만약 Application 실행시 실행은 되지만 에러표시가 난다거나
실행이 안되면서 에러 표시가 나면
Task의 foreign keys에 해당하는 키의 속성을 삭제해주시면 됩니다.

+++ src/main/java/com/otl/otl/domain/Task.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 public class Task {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int tno; // Task 고유 식발자
+    private Long tno; // Task 고유 식발자
 
     @ManyToOne
     @JoinColumn(name = "sno")
